+++ examples/one-month/index.js
@@ -326,16 +326,46 @@ function mainOuterSlot(vido, props) {
     </div>${content}<div class=${overlay}>${loading}</div>`;
 }
 
+/**
+ * @type {import('../../dist/plugins/item-movement').Options}
+ */
+const itemMovementOptions = {
+  // snapToTime: {
+  //   start({ startTime }) {
+  //     return startTime;
+  //   },
+  //   end({ endTime }) {
+  //     return endTime;
+  //   },
+  // },
+  // events: {
+  //   onEnd({ items }) {
+  //     return items.after.map((item) => {
+  //       const initialItem = items.initial.find((currentItem) => currentItem.id === item.id);
+  //       if (!initialItem) return item;
+  //       const initialDurationMs = initialItem.time.end - initialItem.time.start;
+  //       item.time.start = gstc.api.time.date(item.time.start).startOf('day').valueOf();
+  //       item.time.end = gstc.api.time
+  //         .date(item.time.start)
+  //         .add(initialDurationMs, 'millisecond')
+  //         .endOf('day')
+  //         .valueOf();
+  //       return item;
+  //     });
+  //   },
+  // },
+};
+
 const config = {
   licenseKey:
     '====BEGIN LICENSE KEY====\nXOfH/lnVASM6et4Co473t9jPIvhmQ/l0X3Ewog30VudX6GVkOB0n3oDx42NtADJ8HjYrhfXKSNu5EMRb5KzCLvMt/pu7xugjbvpyI1glE7Ha6E5VZwRpb4AC8T1KBF67FKAgaI7YFeOtPFROSCKrW5la38jbE5fo+q2N6wAfEti8la2ie6/7U2V+SdJPqkm/mLY/JBHdvDHoUduwe4zgqBUYLTNUgX6aKdlhpZPuHfj2SMeB/tcTJfH48rN1mgGkNkAT9ovROwI7ReLrdlHrHmJ1UwZZnAfxAC3ftIjgTEHsd/f+JrjW6t+kL6Ef1tT1eQ2DPFLJlhluTD91AsZMUg==||U2FsdGVkX1/SWWqU9YmxtM0T6Nm5mClKwqTaoF9wgZd9rNw2xs4hnY8Ilv8DZtFyNt92xym3eB6WA605N5llLm0D68EQtU9ci1rTEDopZ1ODzcqtTVSoFEloNPFSfW6LTIC9+2LSVBeeHXoLEQiLYHWihHu10Xll3KsH9iBObDACDm1PT7IV4uWvNpNeuKJc\npY3C5SG+3sHRX1aeMnHlKLhaIsOdw2IexjvMqocVpfRpX4wnsabNA0VJ3k95zUPS3vTtSegeDhwbl6j+/FZcGk9i+gAy6LuetlKuARjPYn2LH5Be3Ah+ggSBPlxf3JW9rtWNdUoFByHTcFlhzlU9HnpnBUrgcVMhCQ7SAjN9h2NMGmCr10Rn4OE0WtelNqYVig7KmENaPvFT+k2I0cYZ4KWwxxsQNKbjEAxJxrzK4HkaczCvyQbzj4Ppxx/0q+Cns44OeyWcwYD/vSaJm4Kptwpr+L4y5BoSO/WeqhSUQQ85nvOhtE0pSH/ZXYo3pqjPdQRfNm6NFeBl2lwTmZUEuw==\n====END LICENSE KEY====',
-  innerHeight: 500,
+  innerHeight: 800,
   plugins: [
     HighlightWeekends(),
     TimelinePointer(), // timeline pointer must go first before selection, resizing and movement
     Selection(),
     ItemResizing(), // resizing must fo before movement
-    ItemMovement(),
+    ItemMovement(itemMovementOptions),
     CalendarScroll(),
     ProgressBar(),
     TimeBookmarks({